feat: 계산기 메뉴, 연산식 입력 위치 변경
