+++ src/main/java/com/devcourse/calc/Calculator.java
@@ -4,8 +4,6 @@ import com.devcourse.calc.model.*;
 import com.devcourse.calc.model.Number;
 import com.devcourse.calc.repo.CalcHistoryRepository;
 import com.devcourse.util.Converter;
-import com.devcourse.view.Input;
-import com.devcourse.view.Output;
 
 import java.util.List;
 import java.util.Stack;
@@ -20,28 +18,19 @@ public class Calculator {
         this.converter = converter;
     }
 
-    public void run() {
-        int menu = selectMenu();
-        Output.viewResult(Menu.doAction(menu, this));
+    public String run(int menu) {
+        return Menu.doAction(menu, this);
     }
 
     public String showHistory() {
         return repository.getAllHistories();
     }
 
-    public String calc() {
-        String formula = Input.getFormula();
+    public Number calculate(String formula) {
         List<Token> tokens = converter.infixToPostfixFormula(formula);
-        String result = calculate(tokens).toString();
-        repository.saveHistory(new History(formula, result));
+        Number result = calculate(tokens);
+        repository.saveHistory(new History(formula, result.getNumber()));
         return result;
-    }
-
-    private static int selectMenu() {
-        Output.init();
-        int menu = Input.selectMenu();
-        Output.blankLine();
-        return menu;
     }
 
     private Number calculate(List<Token> tokens) {
@@ -54,7 +43,8 @@ public class Calculator {
 
             Integer firstNumber = stack.pop();
             Integer secondNumber = stack.pop();
-            stack.push(((Operator) token).doAction(secondNumber, firstNumber));
+            Operator operator = (Operator) token;
+            stack.push(operator.execute(secondNumber, firstNumber));
         }
         return new Number(stack.pop());
     }
