+++ src/main/java/com/devcourse/calc/model/Menu.java
@@ -1,6 +1,7 @@
 package com.devcourse.calc.model;
 
 import com.devcourse.calc.Calculator;
+import com.devcourse.view.Input;
 
 import java.util.Arrays;
 import java.util.function.Function;
@@ -8,16 +9,15 @@ import java.util.function.Function;
 public enum Menu {
     NONE(0, "default do not use zero number", calculator -> {throw new RuntimeException("없는 메뉴");}),
     HISTORY(1, "조회", Calculator::showHistory),
-    CALC(2, "계산", Calculator::calc),
-    ;
+    CALC(2, "계산", calculator -> calculator.calculate(Input.getFormula()));
 
     private static final String TO_STRING_TEMPLATE = "%d. %s\n";
 
     private final int number;
     private final String description;
-    private final Function<Calculator, String> action;
+    private final Function<Calculator, Object> action;
 
-    Menu(int number, String description, Function<Calculator, String> action) {
+    Menu(int number, String description, Function<Calculator, Object> action) {
         this.number = number;
         this.description = description;
         this.action = action;
@@ -28,7 +28,7 @@ public enum Menu {
                 .filter(menu -> menu.number == selectedNumber)
                 .findFirst()
                 .orElse(NONE);
-        return selected.action.apply(calculator);
+        return selected.action.apply(calculator).toString();
     }
 
     @Override
